7.7 Dragging notes to empty lanes
It all works! And we were able to get rid of the detachFromLane and
replace it with a more general removeNote.

Do note, though, how inefficient a method removeNote is. It does an
`indexOf` on every single Lane, which is O(n) efficient, and then it
uses `indexOf` again on the found lane. With lots of notes and lanes,
this would get very slow! But... who would have so many notes/lanes in
their kanban app that this would be a problem? That seems like a human
problem, too, at that point.